+++ app/routes/components/Button/changelog.codegen.ts
@@ -16,4 +16,5 @@ export const changelog = ''
   + '- 2025-10-08: Add `light` variant, update `ghost` variant styles.\n'
   + '- 2025-10-15: Support change styles when `disabled`.\n'
   + '- 2025-10-20: Support the prop `color`.\n'
+  + '- 2025-10-28: Add `secondary` variant.\n'
   ;