+++ app/routes/components/Button/index.tsx
@@ -1,13 +1,16 @@
-import { ComponentTemplate } from "@/components/ComponentTemplate";
-import { Section } from "@/components/Section";
+import { useState } from "react";
 import { Typography } from "@/components/Typography";
-
+import { DocLayout } from "@/layouts/DocLayout";
 import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
 import Demo01 from "./demos/Demo01";
 import demo01sourceCode from "./demos/Demo01.source_code.codegen";
-import Example01 from "./demos/Example01";
-import example01SourceCode from "./demos/Example01.source_code.codegen";
+import Demo02 from "./demos/Demo02";
+import demo02SourceCode from "./demos/Demo02.source_code.codegen";
+import Demo03 from "./demos/Demo03";
+import demo03SourceCode from "./demos/Demo03.source_code.codegen";
+import Demo04 from "./demos/Demo04";
+import demo04SourceCode from "./demos/Demo04.source_code.codegen";
 import { sourceCode } from "./source_code.codegen";
 
 export function meta(_: Route.MetaArgs) {
@@ -18,27 +21,45 @@ export function meta(_: Route.MetaArgs) {
 }
 
 export default function ButtonPage() {
+  const [tabsValue, setTabsValue] = useState("doc");
+
+  const tabs = [
+    { name: "doc", label: "Document" },
+    { name: "source", label: "Source Code" },
+    { name: "changelog", label: "Changelog" },
+  ];
+
   return (
-    <ComponentTemplate component="Button">
-      <Section.Root title="Button">
-        <Section.Demo node={<Demo01 />} code={demo01sourceCode} />
-        <Typography.H2>Example</Typography.H2>
-        <Typography.H3>With different colors</Typography.H3>
-        <Section.Demo node={<Example01 />} code={example01SourceCode} />
-        <Typography.H2>Usage</Typography.H2>
-        <Typography.P>
-          The Button component is a sample button component. It has several
-          variants:
-        </Typography.P>
-        <Typography.Ul>
-          <Typography.Li>default.</Typography.Li>
-          <Typography.Li>filled.</Typography.Li>
-          <Typography.Li>light.</Typography.Li>
-          <Typography.Li>ghost.</Typography.Li>
-        </Typography.Ul>
-        <Section.SourceCode files={sourceCode} />
-        <Section.Changelog changelog={changelog} />
-      </Section.Root>
-    </ComponentTemplate>
+    <DocLayout>
+      <DocLayout.Title
+        title="Button"
+        desc="Click, click, click."
+        tabsValue={tabsValue}
+        tabs={tabs}
+        onTabsValueChange={setTabsValue}
+      />
+      {tabsValue === "doc" && (
+        <DocLayout.Content>
+          <DocLayout.Live node={<Demo01 />} code={demo01sourceCode} />
+          <Typography.H2>Examples</Typography.H2>
+          <Typography.H3>With different colors</Typography.H3>
+          <DocLayout.Live node={<Demo02 />} code={demo02SourceCode} />
+          <Typography.H3>With different sizes</Typography.H3>
+          <DocLayout.Live node={<Demo03 />} code={demo03SourceCode} />
+          <Typography.H3>With different variants</Typography.H3>
+          <DocLayout.Live node={<Demo04 />} code={demo04SourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "source" && (
+        <DocLayout.Content>
+          <DocLayout.Files files={sourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "changelog" && (
+        <DocLayout.Content>
+          <DocLayout.Changelog changelog={changelog} />
+        </DocLayout.Content>
+      )}
+    </DocLayout>
   );
 }