+++ src/components/Button/index.ts
@@ -1,4 +1,4 @@
-// Button from copy-ui @ 2025-10-20.
+// Button from copy-ui @ 2025-10-28.
 
 export type { ButtonProps } from "./Button";
 export { Button } from "./Button";